contract-validator/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# NLP
spacy==3.7.2
sentence-transformers==2.3.1

# Vector Database
httpx==0.26.0
aiohttp==3.9.1

# AI/ML
openai==1.10.0
tiktoken==0.5.2

# Database
aiosqlite==0.19.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
